Paul v2
A complete re-write of the back end, with a slight change in the exit
keyword in the front-end. The brain has been re-written to be a
thousand times more powerful, and as such, all new systems have been
written. The old brain is still available if need be, perhaps for the
older systems, but it is far inferior.