Speaking
Paul can now speak, based on a toggle in user_info called NOISY. Fixed
a bug in the word tagging that didn't respect the VERBOSE toggle. There
is a bug in OS X 10.8.4 where there is some speech error if using a
'fancy voice', hopefully this will be fixed soon: I cannot hide it.
Quick change to make sure the old system is still supported with a
commented out line in cli.py. Finder is no longer using "WH" words as a
search item to try and fix a bug, but probably affects results. It
remains to be seen as to which is worse.