Verbs, Verbosity and "it"
Better decision handling if there are no known nouns in the request,
but only in supported modules. Finder now asks what you want to open if
there are a few results. When in verbose mode, more description is
given, and some useless parts have been removed. If you want to
interact with the user, there is a new interact function, which is
capable of returning what item a user chose from a list of 5 or less
items (1 to 5, not 0 to 4, so take off 1). It still works if the user
inputs 'first' or 'third' or the like. As you can probably tell, I use
the tangled movie for testing how well finder can locate stuff.

NEW FEATURE: "it" support. E.g. you tell paul to find something, but
you meant to say open. Just say "open it", and he will open the file he
was told to find a second ago. Or ask what the weather is, but what is
that in Fahrenheit? Just ask "what is that in Fahrenheit?", and Paul
will show you wolfram alpha with the conversion.