Documentation
This is primarily a documentation push. A small change to some of
brain's functions that shouldn't really be noticed, a small bug fix,
and loader.py has been made into a Paul module, which means you can ask
Paul to reload your modules for you. README has been extended to
include a Module How-to.