Object Sentence
The sentence is now an object. This makes everything much more
flexible, and more logical. For example, to get the verbs in the object
'sentence', just use 'sentence.get_parts("VB")'. You can still do
things like iterating over the sentence as you could before, but many
of the functions before are now methods. Most notably, the get_parts
method (as above) and the group_together method are no longer functions
that need the sentence passed into them. group_together has been
renamed keywords, but the old name is kept for backward compatibility
(and if it makes more sense, i.e. it isn't for keywords - just ignore
the warning in that case).