Major Rewrite of Decision Making
This update is quite large. Brain has a new way of deciding which
module gets to deal with each sentence: it examines each word, and if
it is associated with a module, that module gets a 'point'. If the word
is associated with lots of modules, they all get a point. After all the
words have been analysed, the module with the most points 'wins', and
that module gets the sentence to process. Also, 'it' is no longer
subbed out by default (in case it was being used more abstractly, e.g.
"Isn't it nice today?"), but it is easily substituted with a call to
'sentence.replace_it()'. All the built-in modules have been updated to
support the new features, along with some other minor changes:
wikipedia's regex has been improved, as has its error messages; weather
now stores it's source in 'it' rather than the temperature, and
correctly gets tomorrows weather when asked; finder can handle requests
better due to the ignore feature in brain2; and loader is smarter about
.py files being the only loaded files. brain2 gets an improved interact
function, as in it now works. vocab has gotten a little smarter to deal
with the new brain2 decision system, and has learned the verb "go".