Logging and Clocks
Two major changes this time: a new clock.py module for time and date
queries; and logging has been improved with a new function in brain2
called log. It will print the info to the screen if the verbose flag is
on in user_info, and it will always update log.txt. The old brain has
been changed to brain_old to reflect it's status as no longer the
default brain, with plans to change brain2 to just brain. The old word
association system has been completely purged in favour of the new
system. The keywords method now accepts an include parameter, where you
pass in a type of word that is not already counted towards the keywords
(for example, subject nouns are normally excluded, but you may want
them included).