Client and Server
A basic PAUL server is here. The idea is that you can use PAUL on a
server for music, information, etc., while the client is on a small,
less powerful computer such as a Raspberry Pi or Phone or something. I
hope there will be a way to return some sort of instruction system
soon, so execute on the client.