Server Interaction Improvements
The brain2.interact method has been improved to work with the new
server setup. There is no updates needed to modules, anything that was
following recommended practice before will work smoothly with the
server, including asked questions. All actions still occur on the
server, however. While this is good for music controls and the like, it
isn't great for finding local files.