Finder Improvements
Finder now searches based on date as well as name. There are also other
minor changes, such as the brain2.replace_it() function now returns
True or False, depending on whether a sustitution occured, and log.txt
has been added to .gitignore so it isn't added (It is of little use on
here really, it was just more convenient for me).