Refactoring
This is the beginnings of a huge refactoring, and work towards
something approaching a sensible API. This will be slow, buggy, and
WILL break code along the way, but it should lead to a better
experience in the future. "paul.py" will be this API, hopefully one day
being the only dependency for paul modules (excluding ones not specific
to Paul himself, e.g. urllib).