0.1 - Restructure and Convert
We have a 0.1 release! PAUL has reached a point where I would say he is
now stable enough to be an alpha (which is a rather dubious milestone).
The code has been restructured to be more 'regular' in layout, so there
is now a proper directory structure. Modules now only need to import
paul to access the API. The log function has been improved to support
an arbitrary number of arguments, which it prints seperated by a space.
All modules now make use of the new API, the old one is gone. PAUL is
now run using ./bin/PAUL while in the project root. This points to
'main.py' in the PAUL directory, which is itself a symlink to cli.py
(which takes command line arguments). It is like this so that if I need
to change how Paul is interacted with, I just change the main.py link
to the new interface, say gui.py.