0.1.1 - Settings, bugfixes, new API functions
The big change here is Settings: Paul can now turn on or off his
talking and verbosity (both require restarts, but Paul can already
handle this). There was also a few new functions added to paul.py: an
iterable function, returning whether an object is a list, tuple or
dist; a trim_word function that removes a word from a list; a has_word
function that checks if a list has a word in it; and join_lists which
takes all the lists you want joined as parameters, and joins them
without dying if one is None or some other unexpected value. Between
all this, there has been some code tidy-up, which is part of the
continuing refactoring, as well as bugfixes for bugs introduced into
the code the last few updates.