0.1.2 - Bugfixes and Logging
A bug that had completely missed my attention was that the fallfack for
an unknown sentence, the discover module, was broken inside brain. It
now successfully catches unknown queries and googles it. There are two
new flags, LOGGING to determine if the log should be written to, and
MAX_LOG_SIZE, that constricts the size of the log file. By default,
this is 2000 lines, more than enough for standard use. Paul can change
LOGGING himself, but it requires a hand-edit to change the log size. (I
also changed user_info name to my name -- I was getting sick of
'Master')