0.1.3 - Search Engines and Cruft Cleaning
Multiple seach engines are now available in the discover module, and
are set in user_info. The forward function is back, this time in Paul,
but is slightly different - It no longer takes a keyword, but a module
name, so only feed it the name of a few core modules (basically
whatever is included by default). log_unknown was a bit of old cruft
that should have been removed long ago. 'group_together' has been
removed, so please update everything to use 'keywords'. Finally,
interact has been updated with another response option: 'arb', for an
arbitrary response. This basically returns the string that was given
back, and can be converted into a sentence using
paul.Sentence(response), if desired.