0.2.0 - Personalities and Servers
Big update here, hence the bump to version 0.2.0. Amid the usual bug
fixes (although this update will introduce a few), the biggest change
is a new module called personality.py that will be extended along with
the rest of them. This gives Paul a bit of life. At the moment, it just
means he can respond to you saying hi better, but that will eventually
extend to him being a bit more conversational. The server has also got
much better, as now it can use the client program to create effects on
the user's computer, not just on the server. This is done thanks to the
new run_script function inside paul.py, which takes any script that
would normally be run outside of normal execution (such as opening the
web browser, or telling iTunes to play), and tells the script where to
be executed, depending on what mode Paul is in - server or local. One
thing to note is that Paul CANNOT be restarted from the clients. The
client program has also become more useful if you've forgotten the
syntax for connecting, just run the script and it will tell you the
syntax. The input prompts have been standardized throughout Paul now
and is changed in user_info for local settings, and in client for
remote settings.