0.2.1 - Wiki fix, Chained instructions
The wikipedia fix makes the module a little more stable, and more
accurately scrapes the first paragraph. When a disambiguation page is
returned, Paul opens the page for you so you can go after what you
want. There is now also support for chained commands, such as "Find the
movie Django Unchained, and open it." Paul will treat this as two
instructions - he will find the movie django unchained, then he will
open 'it', so hopefully the module is putting relevant information into
the 'it' global.