0.2.2 - Server fixes
The server was quite dangerously broken earlier - you may need to ask
Paul to fix your name. Now, it should work much better. It is starting
to become multi-user, to make it actually useful. You can now have
multiple people talk to him, just not all at the same time, or he'll
give the response of everyone's queries to the same person. His
variables such as it are global - this is by design, so multiple people
can hold a conversation with him, e.g. one person asks for some
research, another wants the full details so asks him to open 'that'. He
is designed for small deployments such as around the home, do keep that
in mind.