0.2.3 - Pretty Paul and Bugs
This update consists of mostly bugfixes, as well as a new title that
shows when Paul starts running. The command to run paul is now all in
lowercase letters, 'paul', rather than 'PAUL'. There is a new method to
set the 'it' value from your module: paul.set_it(value). This is nicer
than asking you to use paul.user_info.info['it'] = value. The weather
has hopefully been tidied up a little bit. You can no longer change
Paul's name. You can now set a title for Paul to use when addressing
you. The new install file has not been tested, if you want to try it do
so with extreme caution and a good backup. Nothing should go wrong, but
as always when using sudo, it could.