0.2.4 - Weather Rewrite and Decision Making
The main thing here is the weather rewrite. It was one of the first
modules written, and has been extended over time. It has always been
the testing bed for all the new APIs and stuff. As such, weather got
treated to a rewrite that was desperately needed. The other change,
which is surprisingly small in terms of code, is how the module is
selected. This will have to be watched carefully, as any change to this
system is delecate. Basically, the more common a word is, the less
weight it has, which should be more useful. It is also easier than
having a complicated weighting system.