0.2.5 - Rewrites and Notifications
This update primarily delivers the same love and care that weather got
in the last update (although the only rewrites close to being as
extensive were Finder and Settings). There is a new send_notification
method that will send the user a notification via notification center,
very easy. There is also a filter_out method, that removes any item in
the rest of the lists from the main list. Finally, a change to the
brain, we now send everything to the decision making system, not just
imperative and interrogative statements. If this proves a pain, it may
be reverted.