0.2.6 - Numbers and Strings
This update has several main features: number parsing, documentation,
program-renaming, and improved install.
paul.parse_number(string) is a new function that will take a string,
such as “one hundred and forty two”, and return an integer, in this
case 142. This could be very useful in a huge range of places, and
currently is used to determine list selections inside the paul.interact
function. The documentation update is a minor docstring tweak. This
really does need to be redone properly, but I really don’t have the
time amongst all the other necessary things to do. client and server
have been renamed to paul-client and paul-server, so it makes more
sense when calling them what they actually are. Install has been
updated to be more explicit when something fails. Lastly, a small
tweak: run_script will now correctly run python3 code, and ruby and
perl have been added to the list of languages available. I feel like
this update took a long time for little reward, but hopefully it was
worth it.