0.3.0a - Bug Fix
This is a very tiny update that should have come out weeks ago. Now it
is the summer holidays here in the southern hemisphere I’ve had to get
a job, so I’m extra busy, with no time to work on Paul. This update
adds mute, and fixes a bug in the media module, so now Paul doesn’t do
anything odd if he can’t find your song. I’m working on a bigger
update, may be a long way off, and it may require some module rewrites.