0.4.1 - Bug Fixes and a dash of Personality
This is the first 0.4.x update, so it is primarily bug fixes. One new
thing is a new function: paul.open_URL(url). It does what it says on
the tin. One other new thing is that Paul has gained a little bit more
personality, so has some more comments, and delivers the weather with a
bit more panache. Yes, panache. As usual, it is weather that is being
used as a testing ground for this sort of thing. Finally, the GUI has
been reworked to be a smidge more user-friendly. Hopefully.