0.4.2 - New Users and New Coders
This update introduces a much better interface for a new user. It does
feel a bit like 20 questions, but it is a start. There are plans in the
future to have location automatically detected, but we’re not there
yet. Also new, is someone else to help me actually write this program.
A fresh pair of eyes is doing well at the moment, so I’m thrilled to
have him on board. There is the beginnings of a tutorial section to
help out new users, but it is still far from ready for prime time.
Finally, a medley of bug fixes: again, to weather; a slightly better
DOM object; checking of response argument validity in paul.interact;
Better spelling and punctuation. Enjoy!