0.4.3 - Bug Fixes, New Flags
A small but very necessary update. He has new flags: -nw to run in
command line mode, -qnw to run in command line mode without a splash
screen (good for when he will be used as a simple utility à la ‘bc’)
and -h, help. Also a bit of a personality update, as well as the usual
bug fixes. Particularly critical bugs, in this case. Diagramming is
very broken, so has been disabled for this release. Some updates to
Sentence to make it function again, but I’m not happy with them.