Tutorial and bug fixes
Tutorial has finally been fully implemented, but more safety checks are
necessary in it, as well as in new_user. Sentence now will once again
function correctly when calling has_one_of and has_word - It became
broken in the recent restructuring. As usual with a pre-release, watch
for loads of bugs.