0.4.4 - The Sum of its Parts
The main purpose of this update is to provide a huge update to the math
module, and a bit of polish to emails. However, it also bumps up the
version number, as quite a bit has been achieved since the last update
(and we’ve been running pre-release for long enough!). So, in
actuality, this update includes Roman Numeral support, Email checking,
Tutorials, steps towards platform independence, the ability to run Paul
from anywhere in the OS, and of course, loads of bug fixes and
improvements. One last thing to note, the Password Manager promised
last time has entered private testing. Obviously this is something we
need to get very right. If it passes the private testing, we will
release it for public testing. Lets see how it goes.